Break out email and log into separate modules
This is to set up for future video work
